+++ @utahdts/utah-design-system/react/components/navigation/MenuItem.jsx
@@ -88,7 +88,7 @@ function MenuItem({ currentMenuItem, menuItem }) {
                 )}
                 onClick={() => setIsChildrenOpen((previouslyOpen) => !previouslyOpen)}
                 icon={Icons.IconChevron()}
-                title="expand sub-menu"
+                title="Expand sub-menu"
               />
             )
             : null